footer fix
