+++ .astro/settings.json
@@ -1,5 +1,5 @@
 {
 	"_variables": {
-		"lastUpdateCheck": 1714901512564
+		"lastUpdateCheck": 1727007538877
 	}
 }
