+++ package.json
@@ -11,12 +11,12 @@
     "astro": "astro"
   },
   "dependencies": {
-    "@astrojs/check": "^0.5.10",
-    "@astrojs/mdx": "^2.3.1",
-    "@astrojs/sitemap": "^3.1.4",
+    "@astrojs/check": "^0.9.3",
+    "@astrojs/mdx": "^3.1.7",
+    "@astrojs/sitemap": "^3.1.6",
     "@fontsource/ibm-plex-sans": "^4.5.13",
     "@playform/compress": "^0.0.3",
-    "astro": "^4.7.1",
+    "astro": "^4.15.8",
     "astro-robots-txt": "^1.0.0",
     "sass": "^1.76.0",
     "tailwindcss": "^3.4.3",
