+++ src/components/Footer.astro
@@ -19,7 +19,6 @@ const year = new Date().getFullYear();
         right: 0;
         font-size: 0.8em;
         /* margin-top: 19rem; */
-        position: fixed;
     }
 
     main {
