+++ src/styles/Base.css
@@ -56,8 +56,16 @@ main {
     }
 }
 
+body, html{
+    height: 100%;
+    margin: 0; 
+}
+
 body {
     background-color: #f7f7f7;
+    display: flex;
+    flex-direction: column;
+    min-height: 100vh;
 }
 
 ul, li {
@@ -67,6 +75,11 @@ ul, li {
 a {
     text-decoration: none;
     color: inherit;
+}
+
+footer {
+    padding: 20px;
+    
 }
 
 @media (prefers-color-scheme: dark) {
